fix double time generator
This is only a temporary fix, it must be be implemented a custom
serialization to stop crashing.

+++ DatabaseFunctionsGenerator/Models/ColumnType.cs
@@ -191,7 +191,7 @@ namespace DatabaseFunctionsGenerator
                 case Types.DateTime:
                     return $"DateTime";
                 case Types.Double:
-                    return $"DOUBLE";
+                    return $"Float";
             }
             return "NOT_EXISTING";
         }
